Consume _all_ waiting packets when in process_master.
This was already the case for serial and file reads, but for udp reads
each call to recv returns one UDP packet.  So if we found our recv
call returned valid results, then keep calling recv until we are out of
packets.

This decreases the number of redundant calls to process_idle.