profile: allow DroneAPI to blacklist some very chatty APM messages
as needed. (For another 10% speedup in the tx_perf_test.py).

Done with optimizing for now - total speedup beyond the native rx
stuff a few days ago is 30%.  Tx path is actually fairly fast in python
I see no need to port to native.  The thing that was killing us was
all the STATUSTEXT and COMMAND_ACK messages sent from the vehicle.