mavproxy_smartcamera: Added Geotag log to SmartCamera Module and Sony A7R
Added a log file to the smart camera module. Currently it is located in /sdcard/log, if the path does not exist the module currently fails to load.
Added support for more sony cameras that failed to work before due to a missing command startRecMode.
Added Initial setting of parameters to Shutter Priority Mode, ISO Auto, Shutter Speed 1/1600. These values can be modified by sending the appropriate commands after the module is loaded, or changing the default values in the python script.