magical: add module
That module is meant to act as a helper for the compass calibration process.
Currently it provides a UI for the calibration, where:

    - a video of examples of rotations the user can apply while calibrating is
      played
    - a geodesic grid is rendered representing the completion mask
    - a 3D model of a quadcopter is rendered and positioned accordingly to the
      attitude messages received from the vehicle
    - a 3D arrow is rendered showing where the magnetic field is currently
      pointing

The modules wxvehicle and glrenderer might be useful for other mavproxy tools
too. They can easily be generalized and put under MAVProxy/modules/lib/ in a
later patch in case that's needed.

Note: the quadcopter 3D model was authored by Ronaldo Silva Pereira - I've just
changed some material parameters.