lib: geodesic_grid: implement section search
That is an implementation based on Ardupilot's AP_GeodesicGrid class
(https://github.com/ArduPilot/ardupilot/blob/master/libraries/AP_Math/AP_GeodesicGrid.cpp).
It's basically an adaptation for Python with the difference that inclusive is
always considered False.

Ideally, the code should be shared, so that the Python implementation would
just wrap over a compiled library. That could be done later.

This should be used for future patches that will augument the compass
calibration process by giving rotations hints to the user.