teach mavproxy about new "joymap" module
This is an alternative to the existing "joystick" module.  Unlike the
joystick module, it loads joystick definitions dynamically from YAML
files, which makes it easier for someone to add support for a device
without needing to patch mavproxy itself.

Additionally, the "joymap" module supports a wider variety of control
types.  A "hat" can be used as a pair of toggle switches, and buttons
can be used either as momentary switches or grouped together to act
like a multi-position switch.