add serial proxy to serial module
this adds support to the 'serial' module for creating a proxy between
a tcp port and a serial connection.  It can be used, for example, to
expose the nsh console on a tcp port.  This adds the 'serial proxy
start', 'serial proxy stop', and 'serial proxy status' commands, and
adds some settings for controlling the proxy.