Add sync function to send timesync message
This is something I need for my own tools. It will send a "timesync" message and print the response to commandline. My tools use this to correlate the drone time with the local time. I have a fork with this change, but proposing to add it if others can use it.