MAVExplorer.py: load graphs before calculating flight mode list
calculating the flightmode_list does a rewind on the messages, which
clears DFReader's message cache, meaning that the expressions don't
parse.

This previously worked because we took a reference to the messages
(mestate.status.msgs), and the rewind assigned over the original hash.
This meant we were holding stale messages, which allowed the expressions
to work.  However, each of the messages held a reference to its "parent"
object (the DFReader), and when [indexed] looked in that parent object
for its "parent" value (so NKF2[1] would go through __getitem__ in
DFMessage, that would look in self_parent.messages["NKF2_1"] - and that
fails because the DFReader has been rewound and assigned over the old
self.messages