MAVExplorer: Fix stats command for DF text and TLOG logs
For DF text: name not message id is used to index counts, and count is
not set for messages that are never seen.

For TLOG: As we don't have a formats attribute on the mavmmaplog class,
we count message instances rather than size in bytes.
Output formatting is adjusted to take account of the longest name.

The filehandle attribute is called f instead in mavmmaplog class, so
updated MPDataLogChildTask class of multiproc_util.py to handle this.