mavflightview.py: correct plotting of GPS data
the "type" variable was being over-written to include the instance number before we dropped down through the code which specially-treats messages based on the type variable.

So stop overwriting the variable before that switch statement, create a special string for when we need it.

Visible difference with this PR is that the GPS plots on the map no longer have lines starting from 0,0 if the GPS starts with no fix.