map: check events in idle loop
currently we only check for events when we receive mavlink packets.

That's a problem for the right-click menu option "Google Maps Link", because often when you want that option you are not actually getting packets from the vehicle....

Process packets in the idle loop as well as after receiving a mavlink packet.